+++ resources/views/emails/withdrawal-approval.blade.php
@@ -3,9 +3,9 @@
 
 Please confirm your withdrawal request.
 - Recipient Address: {{ $target }}
-- Amount: {{ $amount }}
-- Fee: {{ $fee }}
-- **TOTAL**: {{ $fee + $amount }}
+- Amount: {{ $amount }} ADS
+- Fee: {{ $fee }} ADS
+- **TOTAL**: {{ $total }} ADS
 
 @component('mail::button', ['url' => $url])
 Confirm Withdrawal