+++ CHANGELOG.md
@@ -6,11 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Comparing campaign targeting with the schema
+- User referrer
 
 ## [0.9.0] - 2019-05-30
 ### Added
 - Event statistics reports for operators
-- Comparing campaign targeting with the schema
 ### Improved
 - Bonus spending scheme
 ### Fixed